requirements.txt
fastapi
uvicorn
python-jose[cryptography]
passlib[bcrypt]
httpx
duckduckgo-search
google-genai
slowapi
python-multipart
python-dotenv